refactor: rename test files to include the _tests suffix
