--- tests/src/common/bvh/mod.rs
@@ -1,1 +1,0 @@
-mod bvh_tests;
